Pull status message functionality out of the add to cart component and into the product interaction component
