+++ KyleTODO.md
@@ -2,12 +2,7 @@
 
 -- if sale price (found in styles) is not null, strikethrough the original price and display the sale price below the original price
 
--- create size option buttons
-
--- create disabled states for size buttons and style selections
----a style can't be disabled! all styles are options, but sizing options for each style might be disabled if out of stock for that size
-
--- find a way to hold the current skus which will possibly disable sizing options
+---sizing options for each style might be disabled if out of stock for that size
 
 -- fix carousel thumbnails for products with lots of photos, product 40348 is an example. Need to be able to scroll through the photos
 
@@ -16,3 +11,5 @@
 -- handle null styles (40345)
 
 -- add to cart message is staying after switching to new styles
+
+--PRIORITY -- pull 'add to cart' status message out into the product interaction component
